+++ packages/types/src/workspace-config.ts
@@ -22,6 +22,11 @@ export interface CodeownersConfig {
 	 */
 	orderBy: CodeownersOrderBy;
 	/**
+	 * Bitbucket and GitLab only. The number of approvals required for the
+	 * request to be satisfied. This will be applied to all paths.
+	 */
+	requiredApprovals: number | null;
+	/**
 	 * Generates a `CODEOWNERS` file after aggregating all ownership
 	 * rules from each project in the workspace.
 	 */
@@ -330,6 +335,11 @@ export interface PartialCodeownersConfig {
 	 * @default 'file-source'
 	 */
 	orderBy?: CodeownersOrderBy | null;
+	/**
+	 * Bitbucket and GitLab only. The number of approvals required for the
+	 * request to be satisfied. This will be applied to all paths.
+	 */
+	requiredApprovals?: number | null;
 	/**
 	 * Generates a `CODEOWNERS` file after aggregating all ownership
 	 * rules from each project in the workspace.